refactor(deltaboards): Merged deltaboard files, divided large methods and some more cleanup
